feat(react-from-scratch): Using Parcel as the bundler of our app and refactor Pet component
