+++ .gitignore
@@ -1,3 +1,7 @@
 node_modules
 .env
 .DS_Store
+.vscode/
+coverage/
+dist/
+.cache/
