+++ acquire-new-pet/src/App.js
@@ -1,10 +1,6 @@
-const Pet = (props) => {
-  return React.createElement("div", {}, [
-    React.createElement("h1", {}, props.name),
-    React.createElement("h2", {}, props.animal),
-    React.createElement("h2", {}, props.breed),
-  ]);
-}
+import React from 'react'
+import ReactDOM from 'react-dom';
+import Pet from './Pet';
 
 const App = () => {
   return React.createElement("div", {}, [
