+++ acquire-new-pet/src/index.html
@@ -8,8 +8,6 @@
 </head>
 <body>
   <div id="root">not rendered yet.</div>
-  <script src="https://unpkg.com/react@17.0.1/umd/react.development.js"></script>
-  <script src="https://unpkg.com/react-dom@17.0.1/umd/react-dom.development.js"></script>
   <script src="App.js"></script>
 </body>
 </html>
